+++ docs/changelog.rst
@@ -14,7 +14,8 @@ v2.0.0 (in development)
   - **Breaking**: The ``version`` argument passed to ``format`` method
     callables has been renamed to ``base_version``.
 
-- `Versioningit.get_version()` now takes an optional ``write`` argument
+- `Versioningit.get_version()` now takes optional ``write`` and ``fallback``
+  arguments
 
 - The ``onbuild`` step is no longer run when building from an sdist; the
   configuration therefore no longer needs to be idempotent